Add RNASeqGeneExpressionQuantificationInBatch
